+++ x/feegrant/client/cli/tx_test.go
@@ -1,7 +1,6 @@
 package cli_test
 
 import (
-	"bytes"
 	"fmt"
 	"io"
 	"strings"
@@ -14,6 +13,8 @@ import (
 	"github.com/cosmos/gogoproto/proto"
 	"github.com/stretchr/testify/suite"
 
+	_ "cosmossdk.io/api/cosmos/feegrant/v1beta1"
+	_ "cosmossdk.io/api/cosmos/gov/v1beta1"
 	"cosmossdk.io/x/feegrant"
 	"cosmossdk.io/x/feegrant/client/cli"
 	"cosmossdk.io/x/feegrant/module"
@@ -70,7 +71,6 @@ func (s *CLITestSuite) SetupSuite() {
 		WithOutput(io.Discard).
 		WithChainID("test-chain")
 
-	var outBuf bytes.Buffer
 	ctxGen := func() client.Context {
 		bz, _ := s.encCfg.Codec.Marshal(&sdk.TxResponse{})
 		c := clitestutil.NewMockCometRPC(abci.ResponseQuery{
@@ -79,7 +79,7 @@ func (s *CLITestSuite) SetupSuite() {
 
 		return s.baseCtx.WithClient(c)
 	}
-	s.clientCtx = ctxGen().WithOutput(&outBuf)
+	s.clientCtx = ctxGen()
 
 	if testing.Short() {
 		s.T().Skip("skipping test in unit-tests mode.")
@@ -640,7 +640,6 @@ func (s *CLITestSuite) TestFilteredFeeAllowance() {
 	pub, err := k.GetPubKey()
 	s.Require().NoError(err)
 	grantee := sdk.AccAddress(pub.Address())
-
 	clientCtx := s.clientCtx
 
 	commonFlags := []string{
@@ -649,15 +648,12 @@ func (s *CLITestSuite) TestFilteredFeeAllowance() {
 		fmt.Sprintf("--%s=%s", flags.FlagFees, sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdkmath.NewInt(100))).String()),
 	}
 	spendLimit := sdk.NewCoin("stake", sdkmath.NewInt(1000))
-
 	allowMsgs := strings.Join([]string{sdk.MsgTypeURL(&govv1beta1.MsgSubmitProposal{}), sdk.MsgTypeURL(&govv1.MsgVoteWeighted{})}, ",")
 
 	testCases := []struct {
 		name         string
 		args         []string
-		expectErr    bool
-		respType     proto.Message
-		expectedCode uint32
+		expectErrMsg string
 	}{
 		{
 			"invalid granter address",
@@ -671,7 +667,7 @@ func (s *CLITestSuite) TestFilteredFeeAllowance() {
 				},
 				commonFlags...,
 			),
-			true, &sdk.TxResponse{}, 0,
+			"key not found",
 		},
 		{
 			"invalid grantee address",
@@ -685,7 +681,7 @@ func (s *CLITestSuite) TestFilteredFeeAllowance() {
 				},
 				commonFlags...,
 			),
-			true, &sdk.TxResponse{}, 0,
+			"decoding bech32 failed",
 		},
 		{
 			"valid filter fee grant",
@@ -699,7 +695,7 @@ func (s *CLITestSuite) TestFilteredFeeAllowance() {
 				},
 				commonFlags...,
 			),
-			false, &sdk.TxResponse{}, 0,
+			"",
 		},
 	}
 
@@ -709,22 +705,21 @@ func (s *CLITestSuite) TestFilteredFeeAllowance() {
 		s.Run(tc.name, func() {
 			cmd := cli.NewCmdFeeGrant(codecaddress.NewBech32Codec("cosmos"))
 			out, err := clitestutil.ExecTestCLICmd(clientCtx, cmd, tc.args)
-
-			if tc.expectErr {
+			if tc.expectErrMsg != "" {
 				s.Require().Error(err)
+				s.Require().Contains(err.Error(), tc.expectErrMsg)
 			} else {
 				s.Require().NoError(err)
-				s.Require().NoError(clientCtx.Codec.UnmarshalJSON(out.Bytes(), tc.respType), out.String())
+				msg := &sdk.TxResponse{}
+				s.Require().NoError(clientCtx.Codec.UnmarshalJSON(out.Bytes(), msg), out.String())
 			}
 		})
 	}
 
 	// exec filtered fee allowance
 	cases := []struct {
-		name         string
-		malleate     func() error
-		respType     proto.Message
-		expectedCode uint32
+		name     string
+		malleate func() error
 	}{
 		{
 			"valid proposal tx",
@@ -735,8 +730,6 @@ func (s *CLITestSuite) TestFilteredFeeAllowance() {
 					fmt.Sprintf("--%s=%s", flags.FlagFees, sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdkmath.NewInt(100))).String()),
 				)
 			},
-			&sdk.TxResponse{},
-			0,
 		},
 		{
 			"valid weighted_vote tx",
@@ -746,8 +739,6 @@ func (s *CLITestSuite) TestFilteredFeeAllowance() {
 					fmt.Sprintf("--%s=%s", flags.FlagFees, sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdkmath.NewInt(100))).String()),
 				)
 			},
-			&sdk.TxResponse{},
-			2,
 		},
 		{
 			"should fail with unauthorized msgs",
@@ -768,8 +759,6 @@ func (s *CLITestSuite) TestFilteredFeeAllowance() {
 
 				return err
 			},
-			&sdk.TxResponse{},
-			7,
 		},
 	}