+++ x/mint/keeper/genesis_test.go
@@ -59,13 +59,13 @@ func (s *GenesisTestSuite) SetupTest() {
 
 func (s *GenesisTestSuite) TestImportExportGenesis() {
 	genesisState := types.DefaultGenesisState()
-	genesisState.Minter = types.NewMinter(sdk.NewDecWithPrec(20, 2), math.LegacyNewDec(1))
+	genesisState.Minter = types.NewMinter(math.LegacyNewDecWithPrec(20, 2), math.LegacyNewDec(1))
 	genesisState.Params = types.NewParams(
 		"testDenom",
-		sdk.NewDecWithPrec(15, 2),
-		sdk.NewDecWithPrec(22, 2),
-		sdk.NewDecWithPrec(9, 2),
-		sdk.NewDecWithPrec(69, 2),
+		math.LegacyNewDecWithPrec(15, 2),
+		math.LegacyNewDecWithPrec(22, 2),
+		math.LegacyNewDecWithPrec(9, 2),
+		math.LegacyNewDecWithPrec(69, 2),
 		uint64(60*60*8766/5),
 	)